Unified the code which renders the table of vulnerabilities
### What's done:
- added `vulnerabilityTableView` which renders the table of vulnerabilities on `VulnerabilityCollection`, `UserProfile` and `Organization` pages.
- reworked `vulnerabilityCollectionView`.
- deleted `renderVulnerabilityTableForProfileView` and `organizationVulnerabilitiesTab`.
- added many visual improvements to the table of vulnerabilities.

Closes #2746
It's part of #2643
