+++ save-cloud-common/src/commonMain/kotlin/com/saveourtool/save/filters/VulnerabilityFilter.kt
@@ -17,8 +17,8 @@ import kotlinx.serialization.Serializable
  */
 @Serializable
 data class VulnerabilityFilter(
-    val identifierPrefix: String,
-    val statuses: List<VulnerabilityStatus>?,
+    val identifierPrefix: String = "",
+    val statuses: List<VulnerabilityStatus>? = null,
     val isOwner: Boolean = false,
     val tags: Set<String> = emptySet(),
     val minCriticality: Float = 0.0f,
@@ -28,6 +28,8 @@ data class VulnerabilityFilter(
     val organizationName: String? = null,
 ) {
     companion object {
-        val approved = VulnerabilityFilter("", statuses = listOf(VulnerabilityStatus.APPROVED, VulnerabilityStatus.AUTO_APPROVED))
+        val empty = VulnerabilityFilter()
+        val approved = VulnerabilityFilter(statuses = listOf(VulnerabilityStatus.APPROVED, VulnerabilityStatus.AUTO_APPROVED))
+        val pendingReview = VulnerabilityFilter(statuses = listOf(VulnerabilityStatus.PENDING_REVIEW))
     }
 }
