+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/views/userprofile/UserProfileView.kt
@@ -2,8 +2,6 @@
  * View for UserProfile
  */
 
-@file:Suppress("FILE_NAME_MATCH_CLASS")
-
 package com.saveourtool.save.frontend.components.views.userprofile
 
 import com.saveourtool.save.entities.OrganizationDto
@@ -13,6 +11,7 @@ import com.saveourtool.save.frontend.components.inputform.InputTypes
 import com.saveourtool.save.frontend.components.modal.displayModal
 import com.saveourtool.save.frontend.components.modal.mediumTransparentModalStyle
 import com.saveourtool.save.frontend.components.views.contests.tab
+import com.saveourtool.save.frontend.components.views.vuln.vulnerabilityTableView
 import com.saveourtool.save.frontend.externals.fontawesome.*
 import com.saveourtool.save.frontend.utils.*
 import com.saveourtool.save.info.UserInfo
@@ -93,12 +92,12 @@ val userProfileView: FC<UserProfileViewProps> = FC { props ->
 
         // ===================== RIGHT COLUMN =======================================================================
         div {
-            className = ClassName("col-6 mb-4 mt-2")
+            className = ClassName("col-7 mb-4 mt-2")
             props.currentUserInfo?.globalRole?.let { role ->
                 val tabList = if (role.isSuperAdmin() && props.currentUserInfo?.name == user?.name) {
-                    UserProfileTab.values().map { if (it == UserProfileTab.USERS) "${it.name} ($countUsers)" else it.name }
+                    UserProfileTab.entries.map { if (it == UserProfileTab.USERS) "${it.name} ($countUsers)" else it.name }
                 } else {
-                    UserProfileTab.values().filter { it != UserProfileTab.USERS }
+                    UserProfileTab.entries.filter { it != UserProfileTab.USERS }
                         .map { it.name }
                 }
                 tab(selectedMenu.name, tabList, "nav nav-tabs mt-3") { value ->
@@ -109,7 +108,7 @@ val userProfileView: FC<UserProfileViewProps> = FC { props ->
 
             @Suppress("EMPTY_BLOCK_STRUCTURE_ERROR")
             when (selectedMenu) {
-                UserProfileTab.VULNERABILITIES -> renderVulnerabilityTableForProfileView {
+                UserProfileTab.VULNERABILITIES -> vulnerabilityTableView {
                     this.userName = userName
                 }
                 UserProfileTab.USERS -> renderNewUsersTableForProfileView {}
@@ -151,7 +150,7 @@ enum class UserProfileTab {
         override val defaultTab: UserProfileTab = VULNERABILITIES
         override val regexForUrlClassification = "/${FrontendRoutes.VULN_PROFILE}"
         override fun valueOf(elem: String): UserProfileTab = UserProfileTab.valueOf(elem)
-        override fun values(): Array<UserProfileTab> = UserProfileTab.values()
+        override fun values(): Array<UserProfileTab> = entries.toTypedArray()
     }
 }
